o Move compare.py to cmp-cv workflow level
